requirements.txt
Pillow
numpy
matplotlib
scipy
scikit-image
scikit-learn
networkx
tk